+++ src/background/config.ts
@@ -26,8 +26,16 @@ export type Config = {
   disableFadeAnimation: boolean;
 
   disableExtraSpace: boolean;
+  disablePopupAutoClose: boolean;
+  hideProgressBar: boolean;
+  hideVocabOSuccessfulGrade: boolean;
   disable2DReviewing: boolean;
+
   disableJPDBAutoParsing: boolean;
+  gradeButtonsAtBottom: boolean;
+  moveLinksToPopup: boolean;
+  prioritizePopupAboveWord: boolean;
+  prioritizePopupToRightOfWord: boolean;
 
   showPopupKey: Keybind;
   addKey: Keybind;
@@ -53,31 +61,50 @@ export const defaultConfig: Config = {
   contextWidth: 1,
   forqOnMine: true,
 
-  customWordCSS: `/**** Ideas & Options *****
-Remove the slash+asterisks surrounding the code parts below to enable experimental styling ideas
-I personally use but don't want to enable by default or add as full-on options */
-  
+  customWordCSS: `/***** On by default *****
+These are easily editable or deletable styling options that I'm leaving on by default and think others would like too.
+Feel free to delete them or surround the options you don't want with /* and */
+
+/***** Don't change colors for known, blacklisted, or unparsed words *****
+ * This keeps all colors for JPDB.io pages (for 2D reviewing etc.)  */
+
+.jpdb-word.known:not(div.vocabulary-list .jpdb-word.known) { color: inherit }
+.jpdb-word.blacklisted:not(div.vocabulary-list .jpdb-word.blacklisted) { color: inherit }
+.jpdb-word.unparsed:not(div.vocabulary-list .jpdb-word.unparsed) { color: inherit }
+
+/***** Dim known vocab in JPDB *****/
+
+.vocabulary-list .entry:has(.known),
+.vocabulary-list .entry:has(.learning) { opacity: 0.25 }
+
+/***** Black screen for OLED on JPDB *****/
+
+html.dark-mode, html.dark-mode body {
+  background-color: black !important
+}
+
 /***** Hide furigana on known/learning words unless hovering *****/
-/*
+
 .jpdb-word.known:not(:hover) .jpdb-furi { visibility: hidden; }
 .jpdb-word.learning:not(:hover) .jpdb-furi { visibility: hidden; }
 .jpdb-word.due .jpdb-furi { visibility: hidden; }
 .jpdb-word.failed .jpdb-furi { visibility: hidden; }
-*/
 
-/***** ttsu reader dark mode (white text) colors *****/
-/* Change "white" to "black" for light mode (black text) colors */
-/*
-.w-full.h-full .jpdb-word.known { color: white }
-.w-full.h-full .jpdb-word.blacklisted { color:  white }
-.w-full.h-full .jpdb-word.unparsed { color: white }
-*/
 
-/***** Black screen for OLED on JPDB *****/
+/**** Ideas & Options *****
+ * Remove the slash+asterisks surrounding the code parts below to enable experimental styling ideas.
+ * These are all off by default! */
+
+/***** Make parsed-text colors more visible on grayscale displays (super ugly though) *****/
 /*
-html.dark-mode, html.dark-mode body {
-  background-color: black !important
-}
+.jpdb-word.new { color: #18FFFF}
+.jpdb-word.not-in-deck { color: #40C4FF }
+.jpdb-word.known { color: #1B5E20 }
+.jpdb-word.learning { color: #33691E }
+.jpdb-word.due { color: #FFD180 }
+.jpdb-word.failed { color: #FF8A80 }
+.jpdb-word.unparsed { color: #616161 }
+.jpdb-word.blacklisted { color:  #616161 }
 */
 
 /***** E-ink screen new word visibility border - horizontal text version *****/
@@ -93,24 +120,32 @@ html.dark-mode, html.dark-mode body {
 */`,
 
   customPopupCSS: `/* Make review/mining buttons bigger for mobile */
-/*
-button { padding:20px 0; font-size: 14px; flex-grow:1 }
+
+button { padding:20px 10px; font-size: 14px; flex-grow:1 }
 #mine-buttons button { padding: 10px 0; }
 article { max-height: 50vh }
-*/
+
 
 /* Hide never forget and edit buttons */
-/*
-button.edit-add-review,button.never-forget { display:none; }
-*/`,
+
+button.edit-add-review,button.never-forget { display:none; }`,
 
   showPopupOnHover: false,
   touchscreenSupport: false,
   disableFadeAnimation: false,
 
   disableExtraSpace: false,
+  disablePopupAutoClose: false,
+  hideVocabOSuccessfulGrade: false,
+  hideProgressBar: false,
+
   disable2DReviewing: false,
   disableJPDBAutoParsing: false,
+
+  moveLinksToPopup: true,
+  gradeButtonsAtBottom: true,
+  prioritizePopupAboveWord: true,
+  prioritizePopupToRightOfWord: true,
 
   showPopupKey: { key: 'Shift', code: 'ShiftLeft', modifiers: [] },
   addKey: null,
@@ -162,6 +197,7 @@ export function migrateSchema(config: Config) {
 }
 
 export function loadConfig(): Config {
+  console.log('LOADING CONFIG');
   const config = Object.fromEntries(
     Object.entries(defaultConfig).map(([key, defaultValue]) => [key, localStorageGet(key, defaultValue)]),
   ) as Config;
@@ -175,6 +211,7 @@ export function loadConfig(): Config {
     return defaultConfig;
   }
 
+  console.log(config);
   return config;
 }